perf(compile): run fixed-offset formats without interpreting them
Layout.Parse was losing to time.Parse on RFC 3339, which is the one
format the standard library hand-writes a parser for. It was not losing
on the extraction work. A 7-instruction RFC 3339 program spent 44 ns
where the same extractions written straight-line, calling the same
helpers in this package, spent 15 ns. The other 29 ns was the
interpreter: loop overhead, an indirect branch through the opcode
switch, and the covered/end counters maintained on every instruction.

A format whose fields are all fixed-width does not need interpreting.
Its offsets are known when the program is built, so planFast places each
field in a constant slot at the end of Insts and executeFast reads them
in a straight line. That removes four costs at once: no loop, no
indirect branch, no per-instruction coverage arithmetic, and no bounds
checks, since every index is a constant.

Twenty of the thirty-one round-trip formats qualify. The rest carry a
month name, a weekday, a variable-width number, an ISO week or an
ordinal day, and keep the interpreter unchanged.

linux/arm64, Go 1.26.4, benchstat over 12 runs, all within +/-2%:

    Layout_Parse                      44.75n -> 24.68n  -44.84%
    Parser_Cached                     47.05n -> 26.90n  -42.82%
    Layout_Parse_ISODate              26.66n -> 17.73n  -33.48%
    Layout_Parse_Compact              24.66n -> 17.20n  -30.24%
    vs_Stdlib/dateparsa/RFC3339       45.00n -> 24.82n  -44.85%
    vs_Stdlib/dateparsa/SQL_datetime  42.20n -> 24.16n  -42.75%
    vs_Stdlib/dateparsa/ISO_date      26.47n -> 17.89n  -32.41%
    vs_Stdlib/dateparsa/US_slash      26.41n -> 18.00n  -31.86%
    Parse_ISO8601                     149.6n -> 146.5n   -2.11%
    Parse_SQLDatetime                 184.7n -> 183.5n        ~
    Detect_Only                       137.3n -> 159.8n  +16.35%

Against the stdlib on the same format, both sides handed the layout:

    SQL datetime   24.16n vs 89.40n   3.7x
    ISO date       17.89n vs 56.55n   3.2x
    US slash       18.00n vs 54.95n   3.1x
    RFC 3339       24.82n vs 27.86n   1.1x

Zero allocations on every row, unchanged, and TestLayoutParseZeroAlloc
still runs the whole round-trip table plus the extras.

Detect is the one regression and it is the honest price: it plans the
program and never runs it, so it pays and does not collect. Parse, which
does run it, came out flat to slightly faster, and anything reusing the
layout is a third to a half faster. Planning costs about the same as one
execution saves, so it repays on the second row of a column.

Three things this had to get right, none of them free:

Program did not grow. Adding Width and WidthAlt beside a full int
BaseYear took Layout from 208 bytes to 216, which is the next size
class, and measured +25.8% on Parse_ISO8601 and +16 B/op for nothing.
BaseYear is int32 now and the three share the eight bytes BaseYear had,
so Program is still 168 and Layout still 208.
TestProgramFitsItsSizeClass pins it by measuring an allocation, since
unsafe.Sizeof and reflect are both closed off here.

planFast is not a second interpreter. Written with a switch over the
opcodes and a spans array it cost the detection path 10 to 33 percent,
which is the same indirect branch this change exists to remove. It is a
table lookup and a uint64 bitmap now, and the detection cost went away.

The coverage proof moved rather than weakened. The interpreter sums the
widths it reads and compares the sum and the maximum against the input
length, which a gap and an overlap of the same size defeat between them
because they cancel in a sum. planFast proves the fields tile [0, Width)
with a bit per input byte, which cannot cancel, and executeFast then
settles the whole question with one compare. A fast program is checked
more strictly, once, when it is built.

makeTime replaces time.Date when the location is UTC: 5.9 ns against
8.3. The first draft was wrong for every other zone, because time.Date
reads the fields *in* the location and the day-count arithmetic reads
them as UTC, which moves the instant by the offset. Anything not UTC
keeps time.Date, which is the only thing that reads a zone's transitions
correctly. TestMakeTimeMatchesTimeDate walks every day from 1893 to
2110 and the out-of-range values that reach it, February 31st and the
leap second among them.

executeFast reimplements every extraction executeInner already had
right, so it is checked against it rather than reviewed. The slots sit
at the end of Insts precisely so Insts[:N] stays the program the
compiler emitted, which lets the test clear the widths and run the real
interpreter over the real program.
TestFastAgreesWithInterpreter runs 20 layouts through both over every
single-byte mutation of a valid input, and FuzzFastAgreesWithInterpreter
does the same on arbitrary input; 13.4M executions found no divergence.
It earned its keep on the first run, catching that the two paths return
equal-but-distinct Locations for an offset off the 15-minute table.

README numbers are re-measured, and say which machine each table is
from. benchmarks/baseline.txt is deliberately not touched: it is the
darwin/arm64 reference and overwriting it with linux/arm64 numbers would
silently retarget it. That gap is now written down in README rather than
left for somebody to trip over.

+++ dateparser.go
@@ -107,7 +107,7 @@ func parseWithConfig(s string, cfg config) (ParseResult, error) {
 		return ParseResult{}, &ParseError{Input: s, Message: err.Error(), Cause: ErrNoMatch}
 	}
 	if needsBaseYear {
-		program.BaseYear = baseYear(cfg)
+		program.BaseYear = int32(baseYear(cfg))
 	}
 	t, err := program.Execute(s)
 	if err != nil {
@@ -163,7 +163,7 @@ func Detect(s string, opts ...Option) (*Layout, error) {
 		return nil, &ParseError{Input: s, Message: err.Error(), Cause: ErrNoMatch}
 	}
 	if needsBaseYear {
-		program.BaseYear = baseYear(cfg)
+		program.BaseYear = int32(baseYear(cfg))
 	}
 	return &Layout{
 		program:  program,
@@ -187,7 +187,7 @@ func buildAmbiguousError(s string, cfg config) error {
 			return &ParseError{Input: s, Message: cerr.Error(), Cause: ErrNoMatch}
 		}
 		if needsBaseYear {
-			prog.BaseYear = baseYear(cfg)
+			prog.BaseYear = int32(baseYear(cfg))
 		}
 		t, err := prog.Execute(s)
 		if err == nil {
@@ -207,7 +207,7 @@ func buildAmbiguousError(s string, cfg config) error {
 			return &ParseError{Input: s, Message: cerr.Error(), Cause: ErrNoMatch}
 		}
 		if needsBaseYear {
-			prog.BaseYear = baseYear(cfg)
+			prog.BaseYear = int32(baseYear(cfg))
 		}
 		t, err := prog.Execute(s)
 		if err == nil {
